refactor: remove MainLayout, change file structure
