+++ src/App.tsx
@@ -1,17 +1,22 @@
 import { FC } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import MainLayout from './components/layout/MainLayout';
+import Header from './components/Header';
+import Footer from './components/Footer';
 import Home from './pages/Home';
 import Login from './pages/Login';
 
 const App: FC = () => {
   return (
-    <MainLayout>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/login" element={<Login />} />
-      </Routes>
-    </MainLayout>
+    <>
+      <Header />
+      <main>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/login" element={<Login />} />
+        </Routes>
+      </main>
+      <Footer />
+    </>
   );
 };
