+++ src/components/Header/Header.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
-import ContentWrapper from '../ContentWrapper';
-import Logo from '../../Logo';
-import NavBar from '../../NavBar';
+import ContentWrapper from '../../layouts/ContentWrapper';
+import Logo from '../Logo';
+import NavBar from '../NavBar';
 import s from './Header.module.scss';
 
 const Header: FC = () => {
