--- src/components/layout/MainLayout/MainLayout.tsx
@@ -1,16 +1,0 @@
-import { FC, PropsWithChildren } from 'react';
-import Header from '../Header';
-import Footer from '../Footer';
-// import s from './MainLayout.module.scss';
-
-const MainLayout: FC<PropsWithChildren> = ({ children }) => {
-  return (
-    <>
-      <Header />
-      <main>{children}</main>
-      <Footer />
-    </>
-  );
-};
-
-export default MainLayout;
