+++ src/layouts/ContentWrapper/ContentWrapper.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables' as *;
+@use '../../styles/variables' as *;
 
 .root {
   width: 90%;
