+++ src/pages/Home/Home.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import ContentWrapper from '../../components/layout/ContentWrapper';
+import ContentWrapper from '../../layouts/ContentWrapper';
 import Counter from '../../components/Counter';
 import s from './Home.module.scss';
